Refactored out first and last name component implemented loading state for Profile and Name
